(feat)영화 추천 기능 완료
추가적인 작업이 필요할 것으로 보입니다.

+++ src/main/java/com/grepp/moodlink/app/model/data/movie/MovieService.java
@@ -50,7 +50,7 @@ public class MovieService {
             }
             Set<Genre> genres = genreMap.entrySet().stream()
                     .map(entry -> genreRepository.findById(Long.valueOf(entry.getKey()))
-                            .orElseGet(() -> genreRepository.save(new Genre(entry.getKey(), entry.getValue())))
+                            .orElseGet(() -> genreRepository.save(new Genre(entry.getKey().longValue(), entry.getValue())))
                     )
                     .collect(Collectors.toSet());
             movie.setGenres(genres);
